--- common/app/common/HttpStatusException.scala
@@ -1,5 +1,0 @@
-package common
-
-case class HttpStatusException(status: Int, statusText: String) extends Exception {
-  override val getMessage: String = s"Bad response: HTTP $status: $statusText"
-}
