--- common/app/common/Logging.scala
@@ -1,7 +1,0 @@
-package common
-
-import play.api.Logger
-
-trait Logging {
-  implicit val log = Logger(getClass)
-}
