--- common/app/services/ConfigAgentDefaults.scala
@@ -1,5 +1,0 @@
-package services
-
-object ConfigAgentDefaults {
-  val contents: String = """{"fronts":{"fashion":{"collections":[]},"uk/business":{"collections":[]},"testcard":{"collections":[]},"breaking-news":{"collections":[]},"lifeandstyle/home-and-garden":{"collections":[]},"stage":{"collections":[]},"duracell-power-me":{"collections":[]},"lifeandstyle":{"collections":[]},"sport/tour-de-france-2014":{"collections":[]},"uk":{"collections":[]},"lifeandstyle/food-and-drink":{"collections":[]},"voluntary-sector-network":{"collections":[]},"world":{"collections":[]},"lifeandstyle/love-and-sex":{"collections":[]},"au/money":{"collections":[]},"uk/commentisfree":{"collections":[]},"au/commentisfree":{"collections":[]},"au/sport":{"collections":[]},"uk/culture":{"collections":[]},"music":{"collections":[]},"us/culture":{"collections":[]},"us":{"collections":[]},"public-leaders-network":{"collections":[]},"uk-news":{"collections":[]},"world/usa":{"collections":[]},"world/asia":{"collections":[]},"observer":{"collections":[]},"uk/across-the-guardian":{"collections":[]},"world/series/us-mental-health-crisis":{"collections":[]},"powering-people":{"collections":[]},"us/commentisfree":{"collections":[]},"reader-events/observer-ideas":{"collections":[]},"hilton-weekends":{"collections":[]},"uk/sport":{"collections":[]},"google-news":{"collections":[]},"us/business":{"collections":[]},"science":{"collections":[]},"us/across-the-guardian":{"collections":[]},"music/glastonbury-2014":{"collections":[]},"french-film-first":{"collections":[]},"us/sustainable-business":{"collections":[]},"au/across-the-guardian":{"collections":[]},"technology":{"collections":[]},"society-professionals":{"collections":[]},"sport/australia-sport":{"collections":[]},"travel":{"collections":[]},"us/money":{"collections":[]},"education":{"collections":[]},"housing-network":{"collections":[]},"big-ideas":{"collections":[]},"us/sport":{"collections":[]},"lifeandstyle/live-better":{"collections":[]},"environment":{"collections":[]},"books":{"collections":[]},"football":{"collections":[]},"au":{"collections":[]},"local-government-network":{"collections":[]},"education/universityguide":{"collections":[]},"au/business":{"collections":[]},"uk/money":{"collections":[]},"business/companies":{"collections":[]},"film":{"collections":[]},"au/culture":{"collections":[]},"reader-events":{"collections":[]}}}"""
-}
