--- common/test/test-env/DEV.properties
@@ -1,10 +1,0 @@
-#test properties file
-
-foo=bar
-
-guardian.page.foo=eight
-guardian.page.bar=hello world
-
-bar=foo
-
-aws.region=eu-west-1
