--- common/test/test-env/DEVINFRA.properties
@@ -1,9 +1,0 @@
-#test properties file
-
-foo=bar
-
-guardian.page.foo=eight
-guardian.page.bar=hello world
-
-bar=foo
-
