--- common/test/test-env/GUDEV.properties
@@ -1,8 +1,0 @@
-#test properties file
-
-foo=bar
-
-guardian.page.foo=eight
-guardian.page.bar=hello world
-
-bar=foo
