+++ facia-tool/app/controllers/StatusController.scala
@@ -1,9 +1,8 @@
 package controllers
 
-import common.ExecutionContexts
 import play.api.mvc.{Action, Controller}
 
-object StatusController extends Controller with ExecutionContexts {
+object StatusController extends Controller {
 
   def healthStatus = Action { request =>
     Ok("Ok.")
