+++ facia-tool/app/switchboard/SwitchManager.scala
