+++ facia-tool/app/util/Futures.scala
@@ -1,11 +1,10 @@
 package util
 
-import common.ExecutionContexts
-
+import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.Future
 import scala.util.{Failure, Success}
 
-object Futures extends ExecutionContexts {
+object Futures {
   implicit class RichFuture[A](future: Future[A]) {
     def mapTry = future.map(Success.apply) recover { case error => Failure(error) }
   }
