+++ facia-tool/test/metrics/DurationMetricTest.scala
@@ -1,13 +1,12 @@
 package metrics
 
 import com.amazonaws.services.cloudwatch.model.StandardUnit
-import common.ExecutionContexts
 import org.scalatest.{FlatSpec, Matchers}
 
 import scala.concurrent.Await
 import scala.concurrent.duration._
 
-class DurationMetricTest extends FlatSpec with Matchers with ExecutionContexts{
+class DurationMetricTest extends FlatSpec with Matchers {
 
   "DurationMetric" should "start off empty" in {
     val durationMetric: DurationMetric = DurationMetric("TestMetric", StandardUnit.Count)
