+++ facia-tool/test/tools/FaciaApiTest.scala
@@ -4,9 +4,8 @@ import com.gu.facia.client.models.{CollectionJson, Trail}
 import com.gu.pandomainauth.model.User
 import org.joda.time.DateTime
 import org.scalatest.{DoNotDiscover, FreeSpec, Matchers}
-import test.ConfiguredTestSuite
 
-@DoNotDiscover class FaciaApiTest extends FreeSpec with Matchers with ConfiguredTestSuite {
+@DoNotDiscover class FaciaApiTest extends FreeSpec with Matchers {
 
   "update the published date only for a new article and retain existing article date" - {
