+++ facia-tool/test/util/EnumeratorsTest.scala
@@ -3,12 +3,12 @@ package util
 import org.scalatest.concurrent.ScalaFutures
 import org.scalatest.{DoNotDiscover, FlatSpec, ShouldMatchers}
 import play.api.libs.iteratee.Iteratee
-import test.ConfiguredTestSuite
 import util.Enumerators._
 
+import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.Future
 
-@DoNotDiscover class EnumeratorsTest extends FlatSpec with ShouldMatchers with ScalaFutures with ConfiguredTestSuite {
+@DoNotDiscover class EnumeratorsTest extends FlatSpec with ShouldMatchers with ScalaFutures {
   "enumerate" should "simply enumerate the list if the function applied lifts the value into a Future" in {
     enumerate(List(1, 2, 3))(Future.successful).run(Iteratee.getChunks).futureValue should equal(List(
       1, 2, 3
