+++ facia-tool/test/util/RichFutureTest.scala
@@ -2,13 +2,12 @@ package util
 
 import org.scalatest.concurrent.ScalaFutures
 import org.scalatest.{DoNotDiscover, FlatSpec, ShouldMatchers}
-import test.ConfiguredTestSuite
 import util.Futures._
 
 import scala.concurrent.Future
 import scala.util.{Failure, Success}
 
-@DoNotDiscover class RichFutureTest extends FlatSpec with ShouldMatchers with ScalaFutures with ConfiguredTestSuite {
+@DoNotDiscover class RichFutureTest extends FlatSpec with ShouldMatchers with ScalaFutures {
   "mapTry" should "transform a failed Future into a Future of Failure of the error" in {
     val error = new RuntimeException("Blargh!")
     Future.failed(error).mapTry.futureValue should equal(Failure(error))
